made a separate module for format conversion of experimental data
