+++ infer_structcol/__init__.py
@@ -1,5 +1,6 @@
 from .main import Spectrum, Sample
-from .io import convert_data, load_spectrum
+from .io import load_spectrum
 from .run_structcol import calc_refl_trans
 from .model import calc_model_spect, log_posterior
 from .inference import find_max_like, run_mcmc
+from .format_converter import convert_data, convert_data_csv
